feat: tile height is a float so a flat tile carries its real height
as an integer a flat tile could only be 0, which left the frontend to
invent what "0" looks like. as a float it carries its own minimal height
(0.1) as data, like every other tile.

data migration runs through `mix data_migrate`, not at startup — handling
data on boot is what makes builds time out.

+++ lib/nebulith/catalog/tile.ex
@@ -8,7 +8,7 @@ defmodule Nebulith.Catalog.Tile do
     field :emoji, :string
     field :color_role, :string
     field :blocking, :boolean, default: false
-    field :height, :integer, default: 0
+    field :height, :float, default: 0.0
     field :category, :string
     field :title, :string
     field :image_url, :string
